Reuse nodes when possible
Unlike the vanilla implementation, the SB queues sometimes don't use
nodes (and instead insert values to existing nodes). In that case, the
unused node is no longer freed, and is preserved for the next insertion
the thread tries.

This reduces the execution time for the more memory intensive bags (faa
+ id).